requirements.txt
fastapi
uvicorn
matplotlib
numpy
pillow
torch
torchvision
scipy
python-dotenv
requests
